Move 'internal/experimental/intf' into 'internal/conns'.
