--- internal/experimental/intf/README.md
@@ -1,5 +1,0 @@
-# Experimental ServiceData Interface
-
-This package contains experimental `ServiceData` interface:
-
-* `ServiceData` is implemented by a structure defined in each service package
